Fix broken fallback bsd sysctl parsing
Handle scenario where sysctl fails when we give it a big list of keys
and some of those keys do not exist (as on latest Mac) and we do not
properly call the keys we want one-by-one.

As part of that, extend the CallExt class to take in an array of switches
instead of just a list that must already be sanitized.